+++ types/ramda/test/nthArg-tests.ts
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+() => {
+  R.nthArg(1)('a', 'b', 'c'); // => 'b'
+  R.nthArg(-1)('a', 'b', 'c'); // => 'c'
+};
