+++ types/ramda/test/of-tests.ts
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+() => {
+  const b: number[][] = R.of([1]); // => [[1]]
+  const c: number[] = R.of(1);
+};
