+++ types/ramda/test/path-tests.ts
@@ -1,0 +1,12 @@
+import * as R from 'ramda';
+
+() => {
+  const testPath = ['x', 0, 'y'];
+  const testObj = { x: [{ y: 2, z: 3 }, { y: 4, z: 5 }] };
+
+  R.path(testPath, testObj); // => 2
+  R.path(testPath)(testObj); // => 2
+
+  R.path(['a', 'b'])({ c: { b: 2 } }); // => undefined
+  R.path(['a', 'b'], { c: { b: 2 } }); // => undefined
+};
